+++ tests/integration/UserModelTest.php
@@ -1,5 +1,6 @@
 <?php namespace JobApis\JobsToMail\Tests\Integration;
 
+use JobApis\JobsToMail\Models\CustomDatabaseNotification;
 use JobApis\JobsToMail\Tests\TestCase;
 use JobApis\JobsToMail\Models\User;
 
@@ -51,9 +52,9 @@ class UserModelTest extends TestCase
 
     public function testUserModelCanGetNotificaitons()
     {
-        $user = User::with('notifications')->first();
-
-        foreach ($user->notifications() as $notification) {
+        $notification = CustomDatabaseNotification::where('notifiable_type', 'user')->first();
+        $user = User::where('id', $notification->notifiable_id)->with('notifications')->first();
+        foreach ($user->notifications as $notification) {
             $this->assertEquals($user->id, $notification->notifiable_id);
         }
     }