+++ changelog.md
@@ -13,7 +13,8 @@ All Notable changes to `jobs-to-mail` will be documented in this file.
 
 ### Changed
 - Switched Jobs Hub advert out for new email to upgrade link.
-- Using `mail.from.address` for admin email universally.
+- Using config `mail.from.address` for admin email universally.
+- Upgraded to Laravel 5.5, PHP 7.1.
 
 ### Removed
 - Dice as its API has been [shut down](https://www.jobapis.com/2017/10/28/dice-job-board-api-shut-down/).